requirements.txt.txt
customtkinter==5.2.0
mysql-connector-python==8.2.0
Pillow==10.0.0
